Iniciando test en vista signIn
